feat: Make oder api ready

+++ server/internal/repository/models.go
@@ -35,7 +35,7 @@ type Order struct {
 	Status      string     `db:"status" json:"status"`
 	TableNumber string     `db:"table_number" json:"table_number"`
 	Note        *string    `db:"note" json:"note"`
-	TotalPrice  float64    `db:"total_price" json:"total_price"`
+	TotalPrice  float32    `db:"total_price" json:"total_price"`
 	CreatedAt   time.Time  `db:"created_at" json:"created_at"`
 	DeliveredAt *time.Time `db:"delivered_at" json:"delivered_at"`
 	UpdatedAt   time.Time  `db:"updated_at" json:"updated_at"`
